Move ActorError to actor/errors package.
