+++ pkg/actors/actors.go
@@ -34,6 +34,7 @@ import (
 	"k8s.io/utils/clock"
 
 	"github.com/dapr/components-contrib/state"
+	actorerrors "github.com/dapr/dapr/pkg/actors/errors"
 	"github.com/dapr/dapr/pkg/actors/internal"
 	"github.com/dapr/dapr/pkg/actors/placement"
 	"github.com/dapr/dapr/pkg/actors/reminders"
@@ -429,7 +430,7 @@ func (a *actorsRuntime) Call(ctx context.Context, req *invokev1.InvokeMethodRequ
 	}
 
 	if err != nil {
-		if resp != nil && IsActorError(err) {
+		if resp != nil && actorerrors.Is(err) {
 			return resp, err
 		}
 
@@ -574,7 +575,7 @@ func (a *actorsRuntime) callLocalActor(ctx context.Context, req *invokev1.Invoke
 
 	// The .NET SDK signifies Actor failure via a header instead of a bad response.
 	if _, ok := resp.Headers()["X-Daprerrorresponseheader"]; ok {
-		return resp, NewActorError(resp)
+		return resp, actorerrors.NewActorError(resp)
 	}
 
 	return resp, nil
@@ -617,7 +618,7 @@ func (a *actorsRuntime) callRemoteActor(
 
 	// Generated gRPC client eats the response when we send
 	if _, ok := invokeResponse.Headers()["X-Daprerrorresponseheader"]; ok {
-		return invokeResponse, teardown, NewActorError(invokeResponse)
+		return invokeResponse, teardown, actorerrors.NewActorError(invokeResponse)
 	}
 
 	return invokeResponse, teardown, nil
